requirements.txt
exceptiongroup==1.3.1
greenlet==3.2.4
iniconfig==2.3.0
packaging==25.0
playwright==1.56.0
pluggy==1.6.0
pyee==13.0.0
Pygments==2.19.2
pytest==9.0.1
python-dotenv==1.2.1
tomli==2.3.0
typing_extensions==4.15.0
